Fix Crash when a Customization TextField is empty
The empty string caused the  String parsing for the textfields
- expectedBaseHeight
- expectedHeightVariation
- actualHeight
to fail with a NumberFormatException, resulting in a crash.
This fix catches that exception, turns the cursor in the field red (and as soon as a proper value is detected back to grey) and if the menu is left with an empty field, that field resets to default.